refactor: endGame
